style(text-list): move directory of `PTextList`
