+++ src/data-display/tables/definition-table/definition/PDefinition.vue
@@ -1,13 +1,24 @@
 <template>
-    <tr class="p-definition" :class="{block}">
-        <td class="key" :class="{'auto-width': autoKeyWidth, 'no-copy-button': disableCopy}">
-            <slot name="key" v-bind="{name, label, data, value: displayData}">
+    <tr class="p-definition"
+        :class="{block}"
+    >
+        <td class="key"
+            :class="{'auto-width': autoKeyWidth, 'no-copy-button': disableCopy}"
+        >
+            <slot name="key"
+                  v-bind="{name, label, data, value: displayData}"
+            >
                 {{ label || name }}
             </slot>
         </td>
-        <td class="value-wrapper" :class="{'auto-width': autoKeyWidth}">
+        <td class="value-wrapper"
+            :class="{'auto-width': autoKeyWidth}"
+        >
             <span class="value">
-                <slot v-if="disableCopy" name="default" v-bind="{name, label, data, value: displayData}">
+                <slot v-if="disableCopy"
+                      name="default"
+                      v-bind="{name, label, data, value: displayData}"
+                >
                     <template v-if="dataType === 'object'">
                         <p-tag v-for="([objKey, objValue], idx) in Object.entries(displayData)"
                                :key="`tag-${idx}-${objKey}`"
@@ -24,11 +35,14 @@
                     </template>
                 </slot>
                 <p-copy-button v-else
-                               width="0.8rem" height="0.8rem"
+                               width="0.8rem"
+                               height="0.8rem"
                                :value="copyValueFormatter ? copyValueFormatter(data, $props) : copyValue"
                                auto-hide-icon
                 >
-                    <slot name="default" v-bind="{name, label, data, value: displayData}">
+                    <slot name="default"
+                          v-bind="{name, label, data, value: displayData}"
+                    >
                         <template v-if="dataType === 'object'">
                             <p-tag v-for="([objKey, objValue], idx) in Object.entries(displayData)"
                                    :key="`tag-${idx}-${objKey}`"
@@ -46,8 +60,12 @@
                     </slot>
                 </p-copy-button>
             </span>
-            <span v-if="$scopedSlots.extra" class="extra">
-                <slot name="extra" v-bind="{name, label, data, value: displayData}" />
+            <span v-if="$scopedSlots.extra"
+                  class="extra"
+            >
+                <slot name="extra"
+                      v-bind="{name, label, data, value: displayData}"
+                />
             </span>
         </td>
     </tr>
@@ -60,8 +78,8 @@ import {
 
 import type { DefinitionProps } from '@/data-display/tables/definition-table/definition/type';
 import PTag from '@/data-display/tags/PTag.vue';
+import PTextList from '@/data-display/text-list/PTextList.vue';
 import PCopyButton from '@/inputs/buttons/copy-button/PCopyButton.vue';
-import PTextList from '@/others/console/text-list/PTextList.vue';
 
 export default defineComponent<DefinitionProps>({
     name: 'PDefinition',
