+++ src/data-display/text-list/PTextList.stories.mdx
@@ -1,6 +1,6 @@
 import { Meta, Canvas, Story, ArgsTable } from '@storybook/addon-docs/blocks';
 import PTextList from './PTextList.vue';
-import { getTextListArgTypes } from '@/others/console/text-list/story-helper';
+import { getTextListArgTypes } from '@/data-display/text-list/story-helper';
 import { ref } from 'vue';
 import { faker } from '@faker-js/faker';
